+++ src/homework1/Person.java
@@ -1,8 +1,8 @@
 package homework1;
 
-import java.util.List;
+//import java.lang.Override;
 
-public abstract class Person implements Researchable {
+public abstract class Person /*implements Researchable*/ {
 
     private String name;
     private String surname;
@@ -21,39 +21,8 @@ public abstract class Person implements Researchable {
     }
     public abstract void printInfo();
 
-    public void printChild(Tree tree) {
-        printResult(this.name, tree, Relation.PARENT);
-    }
-
-    public void printParents(Tree tree) {
-        printResult(this.name, tree, Relation.CHILD);
-    }
-
-    public void printSister(Tree tree) {
-        printResult(this.name, tree, Relation.BROTHER);
-    }
-
-    public void printBrother(Tree tree) {
-        printResult(this.name, tree, Relation.SISTER);
-    }
-
-    private static void printResult(String name, Tree tree, Relation link) {
-        List<Link> links = tree.getLinks();
-        boolean flag = false;
-        for (Link el: links) {
-            if (el.getPersonFirst().getName().equals(name) &&
-                    el.getLink() == link) {
-                el.getPersonSecond().printInfo();
-                flag = true;
-            }
-        }
-        if (!flag) {
-            System.out.println("No info");
-        }
-    }
-
-    @java.lang.Override
-    public java.lang.String toString() {
+    @Override
+    public String toString() {
         return "Person{" +
                 "name='" + name + '\'' +
                 ", surname='" + surname + '\'' +
