+++ src/homework1/Researchable.java
@@ -2,8 +2,9 @@ package homework1;
 
 public interface Researchable {
 
-    void printChild(Tree tree);
-    void printParents(Tree tree);
-    void printSister(Tree tree);
-    void printBrother(Tree tree);
+    void printChild(String name, Tree tree);
+    void printParents(String name, Tree tree);
+    void printSister(String name, Tree tree);
+    void printBrother(String name, Tree tree);
+
 }
